For jimi

+++ exe_cmd.c
@@ -6,7 +6,7 @@
 /*   By: motuomin <motuomin@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/09/11 15:46:38 by motuomin          #+#    #+#             */
-/*   Updated: 2025/01/07 18:23:22 by motuomin         ###   ########.fr       */
+/*   Updated: 2025/01/07 19:54:13 by motuomin         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -76,11 +76,13 @@ int	exe_cmd(t_cmd *cmd, t_ms *ms)
 	if (cmd->infile && (cmd->inredir == INPUT || cmd->inredir == STD_IN))
 		if (!redirect_input(cmd->infile, cmd))
 			return (0);
+	/*
 	if (!cmd->args[0])
 	{
-		unlink(cmd->infile);
+		if (cmd->infile && access(".heredoc_temp"))
+			unlink(cmd->infile);
 		exit(ms->ret_val);
-	}
+	}*/
 	if (cmd->outfile)
 		if (!redirect_output(cmd->outfile, cmd))
 			return (0);
